+++ src/menus/battle_menu.cpp
@@ -1,5 +1,6 @@
 #include "battle_menu.h"
 #include "../impl.h"
+#include "../inventories/inventory_utils.h"
 
 // Refactor this by implementing the structure in question so the code is not ugly
 SISTERRAY_API u32* initializeBattleItemMenuCursor() {
